requirement.txt
asgiref==3.6.0
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
pytz==2023.3
sqlparse==0.4.4
